Refactor peer manager into modular structure
- Split monolithic peer_manager.rs into focused modules:
  - connection.rs: Connection management and limits
  - discovery.rs: Peer discovery and subnet-based selection
  - heartbeat.rs: Periodic heartbeat events and status reporting
  - types.rs: Common types and events
  - mod.rs: Main PeerManager coordination

- Improve code organization and separation of concerns
- Maintain existing functionality while improving maintainability
- Each module has a single responsibility
- Easier to test individual components
- Follows same pattern as existing scoring/ module

This refactor prepares the codebase for future peer blocking
functionality while keeping the current implementation clean
and modular.
